Develop (#529)
* Refactor and remove warnings from select and select range

* Remove warnings from Bootstrap, Auth and Cache

* Add is_array checks where needed

* Refactor the static helper to use non static calls

* Fix static call warning inside the deprecated codebase

* Fix static calls inside deprecated codebase

* Fix warnings in head bundle and refactor admin view helper to non static calls

* Bump to version 3.2.0

* Fix more warnings and update nodejs dependencies

* Extend head testing and fix more warnings

* Extend head testing and fix more warnings

* Fix more warnings part2

* Introducing PHPStan (#522)

* Introducing PHPStan

* Disable PHPStan for non PHP 7.x

* Reduce total warnings to 7

* Reduce total warnings to 2

* Finally warning free

* Fix composer and couple of code inspections

* Reduce Scrutinizer issues part1

* Fix docblocks for fluent interfaces (#523)

* 150 iterator powered filesystem class (#524)

* Introducing the awesome filesystem

* Unit testing for the new filesystem

* Add CallbackFilter to the Filesystem

* Simplify the Hook class similar to Content classes

* Unit testing for the Filesystem part2

* Unit testing for the Filesystem part3

* Unit testing for the Filesystem part4

* Unit testing for the Filesystem part4

* Implement the new Filesystem

* Move Cache to the Filesystem namespace

* Fix directory separator issue on Windows

* Fix directory separator issue on Windows part2

* Fix directory separator issue on Windows part3 and rename Filesystem vars

* Implement the Filesystem everywhere and introduce renderFile() method

* Remove the old Directory class

* Refactor backup and restore to the Filesystem

* 525 split the large admin forms (#526)

* Split article form

* Split article form

* Split up all admin forms

* Move validateFile() method back to the Cache

* Improve PageCache module

* Introducing sorted filesystem array

* Fix minor moduule issues

* Improve the skeleton template

* Finish unit testing for the Filesystem

* Cosmetics on the backup and restore console tasks

* Fix missing whitespace while sending emails

* Fix missing whitespace while sending emails

* Add missing testing for console

* Improve code quality (#528)

* Add missing return statements

* Add missing return statements part2

* Get rid of mixed part1

* Move the Purifier under the Html namespace

* Get rid of mixed part2

* Get rid of mixed part2

* Get rid of mixed part3

* Improve docblock part1

+++ includes/Module/Module.php
@@ -110,7 +110,7 @@ class Module
 
 		/* load the language */
 
-		if (array_key_exists('alias', static::$_moduleArray))
+		if (is_array(static::$_moduleArray) && array_key_exists('alias', static::$_moduleArray))
 		{
 			$this->_language->load(
 			[
@@ -127,12 +127,12 @@ class Module
 	 *
 	 * @param string $type type of the notification
 	 *
-	 * @return mixed
+	 * @return string|array|boolean
 	 */
 
 	public function getNotification($type = null)
 	{
-		if (array_key_exists($type, self::$_notificationArray))
+		if (is_array(self::$_notificationArray) && array_key_exists($type, self::$_notificationArray))
 		{
 			return self::$_notificationArray[$type];
 		}
@@ -149,7 +149,7 @@ class Module
 	 * @since 3.0.0
 	 *
 	 * @param string $type type of the notification
-	 * @param mixed $message message of the notification
+	 * @param string|array $message message of the notification
 	 */
 
 	public function setNotification($type = null, $message = null)
@@ -166,7 +166,7 @@ class Module
 
 	public function install()
 	{
-		if (array_key_exists('alias', static::$_moduleArray))
+		if (is_array(static::$_moduleArray) && array_key_exists('alias', static::$_moduleArray))
 		{
 			$module = Db::forTablePrefix('modules')->create();
 			$module->set(static::$_moduleArray);
@@ -192,7 +192,7 @@ class Module
 
 	public function uninstall()
 	{
-		if (array_key_exists('alias', static::$_moduleArray))
+		if (is_array(static::$_moduleArray) && array_key_exists('alias', static::$_moduleArray))
 		{
 			Db::forTablePrefix('modules')->where('alias', static::$_moduleArray['alias'])->deleteMany();
